requirements.txt
pandas
numpy
matplotlib
requests
pyyaml
dash